Inventory Report-Stock Levels
Added a menu item,  Inventory Reports and sub menu, Stock Levels on the
reports side bar. It is working alright but there is a problem with
including the from and to in the query so i have commented them out on
model Receipt. @kitsao @briankip  have a look at it.